requirements.txt
docker
docker-compose
editdistance
exrex
fire
immutables
mock
orjson
pytest
python-dotenv
pyyaml
requests
streamlit
yamllint
